requirements.txt
fastapi
gpiozero
lgpio
pillow
RPi.GPIO
spidev
waveshare-epaper
